requirements.txt
pandas
altair
pytest
pyyaml
scikit-learn
numpy